refactor(lqip-loader): split test files
