+++ packages/lqip-loader/src/__tests__/utils.test.ts
@@ -10,7 +10,6 @@ import Vibrant from 'node-vibrant';
 import type {Palette} from 'node-vibrant/lib/color';
 
 import {toPalette, toBase64} from '../utils';
-import * as lqip from '../lqip';
 
 describe('lqip-loader', () => {
   describe('toBase64', () => {
@@ -42,26 +41,6 @@ describe('lqip-loader', () => {
 
     it('should return 5 hex colours with no errors if a palette was incomplete', () => {
       expect(toPalette(testSwatchWithNull)).toHaveLength(5);
-    });
-  });
-
-  describe('lqip library', () => {
-    const imgPath = path.join(__dirname, '__fixtures__', 'endi.jpg');
-    const invalidPath = path.join(__dirname, '__fixtures__', 'docusaurus.svg');
-
-    it('should reject unknown or unsupported file format', async () => {
-      await expect(lqip.base64(invalidPath)).rejects.toBeTruthy();
-    });
-
-    it('should generate a valid base64', async () => {
-      const expectedBase64 = 'data:image/jpeg;base64,/9j/2wBDA';
-      await expect(lqip.base64(imgPath)).resolves.toContain(expectedBase64);
-    });
-
-    it('should generate a valid color palette', async () => {
-      const imgPalette = await lqip.palette(imgPath);
-      expect(imgPalette).toHaveLength(6);
-      expect(imgPalette).toContain('#578ca1');
     });
   });
 });
