+++ packages/lqip-loader/src/index.ts
@@ -25,6 +25,8 @@ export default async function lqipLoader(
 
   const config = this.getOptions() || {};
   config.base64 = 'base64' in config ? config.base64 : true;
+  // color palette generation is set to false by default
+  // since it is little bit slower than base64 generation
   config.palette = 'palette' in config ? config.palette : false;
 
   let content = contentBuffer.toString('utf8');
@@ -53,8 +55,6 @@ export default async function lqipLoader(
 
   const outputPromises: [Promise<string> | null, Promise<string[]> | null] = [
     config.base64 === true ? lqip.base64(imgPath) : null,
-    // color palette generation is set to false by default
-    // since it is little bit slower than base64 generation
     config.palette === true ? lqip.palette(imgPath) : null,
   ];
